+++ R/SpectrumList-methods.R
@@ -1,0 +1,6 @@
+#' @include SpectrumList.R
+
+setMethod("show", "SpectrumList", function(object) {
+    .show_SpectrumList(object, margin = "  ", print.classinfo = TRUE)
+})
+