+++ NEWS.md
@@ -1,7 +1,7 @@
 # MSnbase 2.7
 
 ## Changes in version 2.7.8
-- Nothing yet
+- Add SpectrumList class.
 
 ## Changes in version 2.7.7
 - Fix unit tests (issue [#360](https://github.com/lgatto/MSnbase/issues/360), wrong MS OBO CV terms for data smoothing